requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
argparse=1.4.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.10.11=hecd8cb5_0
certifi=2022.12.7=py310hecd8cb5_0
grpcio=1.51.1=pypi_0
grpcio-tools=1.51.1=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
libcxx=14.0.6=h9765a3e_0
libffi=3.4.2=hecd8cb5_6
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
ncurses=6.3=hca72f7f_3
numpy=1.23.5=py310h9638375_0
numpy-base=1.23.5=py310ha98c3c9_0
openssl=1.1.1s=hca72f7f_0
pillow=9.4.0=pypi_0
pip=22.3.1=py310hecd8cb5_0
protobuf=4.21.12=pypi_0
python=3.10.8=h218abb5_1
readline=8.2=hca72f7f_0
setuptools=65.6.3=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h880c91c_0
tk=8.6.12=h5d9f67b_0
tzdata=2022g=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.8=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
